services/server/requirements.txt
django-countries==7.2.1  
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-dotenv==1.4.2
Pillow==9.0.1  
psycopg2-binary==2.9.3
pytest-django==4.5.2  
gunicorn==20.1.0
six==1.13.0
paypal-checkout-serversdk==1.0.1
eth-brownie==1.18.1   